refactor(core): extract seeded coding context lifecycle
